+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added lowering and evaluation of graph `MATCH` label negation, conjunction, and disjunction
+- Added lowering and evaluation of graph `MATCH` `WHERE` clauses
 
 ### Removed